add docker compose, BookCatalog service, update services and fix errors

+++ services/api/src/main/java/com/shadoww/api/dto/response/ChapterResponse.java
@@ -15,7 +15,7 @@ public class ChapterResponse {
 
     long bookId;
 
-    long chapterNumber;
+    int chapterNumber;
 
 //    public ChapterResponse(Chapter chapter) {
 //        this.id = chapter.getId();
